// frontend/src/app/api/proxies/issue/issue.service.ts
interface HttpClientOptions {
  readonly headers?: HttpHeaders | Record<string, string | string[]>;
  readonly context?: HttpContext;
  readonly params?:
        | HttpParams
      | Record<string, string | number | boolean | Array<string | number | boolean>>;
  readonly reportProgress?: boolean;
  readonly withCredentials?: boolean;
  readonly credentials?: RequestCredentials;
  readonly keepalive?: boolean;
  readonly priority?: RequestPriority;
  readonly cache?: RequestCache;
  readonly mode?: RequestMode;
  readonly redirect?: RequestRedirect;
  readonly referrer?: string;
  readonly integrity?: string;
  readonly referrerPolicy?: ReferrerPolicy;
  readonly transferCache?: {includeHeaders?: string[]} | boolean;
  readonly timeout?: number;
}

type HttpClientBodyOptions = HttpClientOptions & {
  readonly observe?: 'body';
};

type HttpClientEventOptions = HttpClientOptions & {
  readonly observe: 'events';
};

type HttpClientResponseOptions = HttpClientOptions & {
  readonly observe: 'response';
};

type HttpClientObserveOptions = HttpClientOptions & {
  readonly observe?: 'body' | 'events' | 'response';
};







@Injectable({ providedIn: 'root' })
export class IssueService {
  private readonly http = inject(HttpClient);
/**
 * @summary List issues for the current org
 */
 getIssues<TData = IssueList>( options?: HttpClientBodyOptions): Observable<TData>;
 getIssues<TData = IssueList>( options?: HttpClientEventOptions): Observable<HttpEvent<TData>>;
 getIssues<TData = IssueList>( options?: HttpClientResponseOptions): Observable<AngularHttpResponse<TData>>;
  getIssues<TData = IssueList>(
     options?: HttpClientObserveOptions): Observable<TData | HttpEvent<TData> | AngularHttpResponse<TData>> {
    if (options?.observe === 'events') {
      return this.http.get<TData>(
      `/api/issues`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'events',
      }
    );
    }

    if (options?.observe === 'response') {
      return this.http.get<TData>(
      `/api/issues`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'response',
      }
    );
    }

    return this.http.get<TData>(
      `/api/issues`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'body',
      }
    );
  }
/**
 * @summary Create an issue
 */
 postIssue<TData = Issue>(newIssue?: NewIssue, options?: HttpClientBodyOptions): Observable<TData>;
 postIssue<TData = Issue>(newIssue?: NewIssue, options?: HttpClientEventOptions): Observable<HttpEvent<TData>>;
 postIssue<TData = Issue>(newIssue?: NewIssue, options?: HttpClientResponseOptions): Observable<AngularHttpResponse<TData>>;
  postIssue<TData = Issue>(
    newIssue?: NewIssue, options?: HttpClientObserveOptions): Observable<TData | HttpEvent<TData> | AngularHttpResponse<TData>> {
    if (options?.observe === 'events') {
      return this.http.post<TData>(
      `/api/issues`,
      newIssue,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'events',
      }
    );
    }

    if (options?.observe === 'response') {
      return this.http.post<TData>(
      `/api/issues`,
      newIssue,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'response',
      }
    );
    }

    return this.http.post<TData>(
      `/api/issues`,
      newIssue,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'body',
      }
    );
  }
/**
 * @summary Update an issue
 */
 updateIssue<TData = Issue>(id: number,
    issuePatch?: IssuePatch, options?: HttpClientBodyOptions): Observable<TData>;
 updateIssue<TData = Issue>(id: number,
    issuePatch?: IssuePatch, options?: HttpClientEventOptions): Observable<HttpEvent<TData>>;
 updateIssue<TData = Issue>(id: number,
    issuePatch?: IssuePatch, options?: HttpClientResponseOptions): Observable<AngularHttpResponse<TData>>;
  updateIssue<TData = Issue>(
    id: number,
    issuePatch?: IssuePatch, options?: HttpClientObserveOptions): Observable<TData | HttpEvent<TData> | AngularHttpResponse<TData>> {
    if (options?.observe === 'events') {
      return this.http.patch<TData>(
      `/api/issues/${id}`,
      issuePatch,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'events',
      }
    );
    }

    if (options?.observe === 'response') {
      return this.http.patch<TData>(
      `/api/issues/${id}`,
      issuePatch,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'response',
      }
    );
    }

    return this.http.patch<TData>(
      `/api/issues/${id}`,
      issuePatch,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'body',
      }
    );
  }
/**
 * @summary Soft-delete an issue
 */
 deleteIssue<TData = void>(id: number, options?: HttpClientBodyOptions): Observable<TData>;
 deleteIssue<TData = void>(id: number, options?: HttpClientEventOptions): Observable<HttpEvent<TData>>;
 deleteIssue<TData = void>(id: number, options?: HttpClientResponseOptions): Observable<AngularHttpResponse<TData>>;
  deleteIssue<TData = void>(
    id: number, options?: HttpClientObserveOptions): Observable<TData | HttpEvent<TData> | AngularHttpResponse<TData>> {
    if (options?.observe === 'events') {
      return this.http.delete<TData>(
      `/api/issues/${id}`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'events',
      }
    );
    }

    if (options?.observe === 'response') {
      return this.http.delete<TData>(
      `/api/issues/${id}`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'response',
      }
    );
    }

    return this.http.delete<TData>(
      `/api/issues/${id}`,{
        ...(options as Omit<NonNullable<typeof options>, 'observe'>),
        observe: 'body',
      }
    );
  }
};