// frontend/src/app/api/proxies/issueTrackerAPI.schemas.ts
export type IssuePriority = typeof IssuePriority[keyof typeof IssuePriority];


export const IssuePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  urgent: 'urgent',
} as const;

export type IssueCategory = typeof IssueCategory[keyof typeof IssueCategory];


export const IssueCategory = {
  bug: 'bug',
  feature_request: 'feature_request',
  ui_ux: 'ui_ux',
  performance: 'performance',
  billing: 'billing',
  other: 'other',
} as const;

export type IssueStatus = typeof IssueStatus[keyof typeof IssueStatus];


export const IssueStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  closed: 'closed',
} as const;

export interface Issue {
  id: number;
  orgId: string;
  createdBy: string;
  description: string;
  /** @nullable */
  assignedUser: string | null;
  /** @nullable */
  assignedUserName?: string | null;
  priority: IssuePriority;
  category: IssueCategory;
  status: IssueStatus;
  /** @nullable */
  screenshotPath: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt: string | null;
}